połączenie widoku wyszukiwarki z dodaniem oferty
