+++ public/locales/en/translations.json
@@ -60,9 +60,7 @@
         "name": "Phone number *",
         "errorEmptyField": "Field can not be empty",
         "errorPhoneRegex": "Please enter a valid phone number e.g. 500500500 or 914646464"
-      }
-    },
-    "offer": {
+      },
       "add": {
         "addOfferText": "Add offer"
       }
