+++ public/locales/pl/translations.json
@@ -81,9 +81,7 @@
         "name": "Telefon *",
         "errorEmptyField": "Pole nie może być puste",
         "errorPhoneRegex": "Podaj poprawny numer telefonu np. 500500500 lub 914646464"
-      }
-    },
-    "offer": {
+      },
       "add": {
         "addOfferText": "Dodaj ofertę"
       }
