+++ src/components/Search/SearchInput.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import SearchForm from 'components/Search/SearchForm/SearchForm';
 import FoundSearchResults from 'components/Search/SearchResults/FoundSearchResults/FoundSearchResults';
 import NoSearchResults from 'components/Search/SearchResults/NoSearchResults/NoSearchResults';
+import OfferAdd from '../UI/Add/OfferAdd';
 
 export default class SearchInput extends React.Component {
   constructor(props) {
@@ -36,6 +37,7 @@ export default class SearchInput extends React.Component {
         <SearchForm
           updateFoundServices={this.updateFoundServices}
         />
+        <OfferAdd />
         {results}
       </div>
     );
