+++ src/components/UI/Add/OfferAdd.css
@@ -6,26 +6,14 @@
   padding: 0;
 }
 
-.row:after {
-  content: '';
-  clear: both;
-  display: table;
-}
-
-.container {
-  margin: 100px auto;
-  max-width: 1000px;
-  padding: 10px 20px;
-}
-
 .add-offer {
   position: relative;
   float: left;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  margin: 100px auto;
   padding: 0 20%;
+  margin-top: 20px;
 }
 
 .add-offer__link {
