+++ src/core/Routes.jsx
@@ -2,15 +2,12 @@ import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import ScreensSearch from 'screens/Search/Search';
 import ScreensAddForm from 'screens/Add/Form/Form';
-import ScreensAddOffer from 'screens/Offer/Add/OfferAdd';
-
 
 export default() => (
   <BrowserRouter>
     <Switch>
       <Route exact path="/" component={ScreensSearch} />
       <Route path="/add/form" component={ScreensAddForm} />
-      <Route path="/add-offer" component={ScreensAddOffer} />
     </Switch>
   </BrowserRouter>
 );
